pyproject.toml
[project]
name = "rag-mgi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepseek-ai>=0.0.1",
    "gradio>=4.12.0",
    "openai>=1.6.1",
    "python-dotenv>=1.0.0",
    "langchain>=0.0.350",
    "langchain-community>=0.0.10",
    "langchain-deepseek>=0.1.0",
    "requests>=2.31.0",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.2.2",
    "pydantic>=2.5.2",
    "python-multipart>=0.0.6",
    "docling>=2.28.4",
    "hf-xet>=1.0.3",
    "langchain-chroma>=0.2.2",
    "langchain-huggingface>=0.1.2",
    "huggingface-hub[hf-xet]>=0.30.1",
]
